+++ src/tb/tb_tempbuffer.cpp
@@ -75,7 +75,7 @@ bool TBTempBuffer::AppendSpace(int size)
 bool TBTempBuffer::AppendString(const char *str)
 {
 	// Add 1 to include the null termination in the data.
-	if (Append(str, strlen(str) + 1))
+	if (Append(str, (int)strlen(str) + 1))
 	{
 		// Now remove the null termination from the append position
 		// again, so another call will append to the same string (instead of
@@ -115,7 +115,7 @@ bool TBTempBuffer::AppendFile(const char *filename)
 {
 	if (TBFile *file = TBFile::Open(filename, TBFile::MODE_READ))
 	{
-		const size_t file_size = file->Size();
+		const int file_size = (int)file->Size();
 		if (Reserve(m_append_pos + file_size + 1) && file->Read(m_data + m_append_pos, 1, file_size) == file_size)
 		{
 			// Increase append position and null terminate
